Fix Pydantic validation error by changing pod_count from int to float to support bound constraints

+++ environment/models.py
@@ -139,7 +139,7 @@ class ServiceHealth(BaseModel):
 
     p99_latency_ms: float
 
-    pod_count: int
+    pod_count: float
 
 class Observation(BaseModel):
